adjust packets for first bunch of tests

+++ Core/Inc/Sensors/GPS/Direction.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-enum class Direction : uint8_t
+enum class Direction : uint16_t
 {
 	NORTH,
 	SOUTH,
